Android: Fixes for Iup official resync

+++ src/android/source/iupandroid_info.c
@@ -78,7 +78,7 @@ int iupdrvGetScreenDepth(void)
 	return 32;
 }
 
-float iupdrvGetScreenDpi(void)
+double iupdrvGetScreenDpi(void)
 {
 #if 0
 	CGRect rect = CGDisplayBounds(kCGDirectMainDisplay);
